Moved style declarations into partial instead of in component
