+++ library/src/scripts/components/headers/pieces/VanillaMobileHomeHeader.tsx
@@ -28,10 +28,8 @@ import VanillaHeaderNavItem from "@library/components/mebox/pieces/VanillaHeader
 import { signIn } from "@library/components/icons";
 import VanillaHeaderListItem from "@library/components/mebox/pieces/VanillaHeaderListItem";
 import { dummyNavigationData } from "@library/components/mebox/state/dummyNavigationData";
-import vanillaHeaderStyles from "@library/components/headers/VanillaHeaderStyles";
 import { style } from "typestyle";
-import { percent, px } from "csx";
-import { layoutStyles } from "@library/styles/layoutStyles";
+import vanillaHeaderClasses from "@library/components/headers/VanillaHeaderStyles";
 
 interface IProps extends IDeviceProps, IInjectableUserState {
     container?: Element; // Element containing header. Should be the default most if not all of the time.
@@ -64,46 +62,20 @@ export class VanillaMobileHomeHeader extends React.Component<IProps> {
             countClass: classNames(countClass, "vanillaHeader-messagesCount"),
         };
 
-        const headerStyles = vanillaHeaderStyles();
-        const mediaQueries = layoutStyles().mediaQueries();
+        const classes = vanillaHeaderClasses();
 
         return ReactDOM.createPortal(
-            <header
-                className={classNames(
-                    "vanillaHeader",
-                    "vanillaHeaderHome",
-                    style({ minHeight: px(parseInt(headerStyles.sizing.mobile.height.toString(), 10) * 2) }),
-                    this.props.className,
-                )}
-            >
+            <header className={classNames("vanillaHeader", "vanillaHeaderHome", classes.root, this.props.className)}>
                 <Container className="vanillaHeaderHome-top">
                     <PanelWidgetHorizontalPadding>
-                        <div
-                            className={classNames(
-                                "vanillaHeader-bar",
-                                "isHome",
-                                style(
-                                    {
-                                        display: "flex",
-                                        justifyContent: "space-between",
-                                        flexWrap: "nowrap",
-                                        alignItems: "center",
-                                        height: headerStyles.sizing.height,
-                                        width: percent(100),
-                                        $nest: {
-                                            "&.isHome": {
-                                                justifyContent: "space-between",
-                                            },
-                                        },
-                                    },
-                                    mediaQueries.oneColumn({ height: headerStyles.sizing.mobile.height }),
-                                ),
-                            )}
-                        >
+                        <div className={classNames("vanillaHeader-bar", "isHome", classes.bar)}>
                             <FlexSpacer className="vanillaHeaderHome-left" />
                             <HeaderLogo
                                 {...dummyLogoData}
-                                className="vanillaHeader-logoContainer"
+                                className={classNames(
+                                    "vanillaHeader-logoContainer",
+                                    style({ display: "inline-flex", alignSelf: "center" }),
+                                )}
                                 logoClassName="vanillaHeader-logo isCentred"
                             />
                             {isGuest ? (
